# swaggerui/swagger.json
{
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "gc-ai API specification\n\nThis should demonstrate all the possible comment annotations\nthat are available to turn go code into a fully compliant swagger 2.0 spec",
    "title": "gc-ai API.",
    "termsOfService": "there are no TOS at this moment, use at your own risk we take no responsibility",
    "contact": {
      "email": "594780735@qq.com"
    },
    "license": {
      "name": "MIT",
      "url": "http://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "host": "47.102.147.41:8080",
  "basePath": "/",
  "paths": {
    "/api/v1/battle": {
      "get": {
        "description": "query the process of a battle",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "queryProcess",
        "parameters": [
          {
            "type": "integer",
            "format": "uint64",
            "x-go-name": "BattleID",
            "name": "battle_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/queryProcessResp"
          },
          "400": {
            "$ref": "#/responses/statusResponse"
          }
        }
      },
      "post": {
        "description": "start a battle",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "startBattle",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/StartBattleReq"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/startBattleResp"
          },
          "400": {
            "$ref": "#/responses/statusResponse"
          }
        }
      }
    },
...
    "/api/v1/sourcecode": {
      "get": {
        "description": "get the user's all source codes",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "getSrcByUserAndGame",
        "parameters": [
          {
            "type": "integer",
            "format": "uint64",
            "x-go-name": "GameID",
            "description": "game id",
            "name": "game_id",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/getSourceCodesResp"
          },
          "400": {
            "$ref": "#/responses/statusResponse"
          }
        }
      },
      "post": {
        "description": "upload source code",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "uploadSourceCode",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UploadSourceCodeReq"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/uploadSourceCodeResp"
          },
          "400": {
            "$ref": "#/responses/statusResponse"
          }
        }
      }
    },
    "/api/v1/user/battles": {
      "get": {
        "description": "get user's battles",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "getUserBattles",
        "responses": {
          "200": {
            "$ref": "#/responses/getUserBattlesResp"
          },
          "400": {
            "$ref": "#/responses/statusResponse"
          }
        }
      }
    },
    "/api/v1/user/sourcecode": {
      "get": {
        "description": "get the user's all source codes",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "getSourceCodeByUser",
        "responses": {
          "200": {
            "$ref": "#/responses/getSourceCodesResp"
          },
          "400": {
            "$ref": "#/responses/statusResponse"
          }
        }
      }
    },
    "/auth": {
      "post": {
        "description": "get jwt token from server",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "getAuth",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UserJSON"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/authSuccess"
          },
          "400": {
            "$ref": "#/responses/statusResponse"
          }
        }
      }
    },
    "/register": {
      "post": {
        "description": "register a new account",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "register",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/RegInfo"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/authSuccess"
          },
          "400": {
            "$ref": "#/responses/statusResponse"
          }
        }
      }
    },
    "/vericode": {
      "post": {
        "description": "request to server to send a verification code",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "sendVeriCode",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/TargetEmail"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/statusResponse"
          },
          "400": {
            "$ref": "#/responses/statusResponse"
          },
          "500": {
            "$ref": "#/responses/statusResponse"
          }
        }
      }
    }
  },
  "definitions": {
...
    "AuthSuccess": {
      "type": "object",
      "properties": {
        "code": {
          "description": "response status",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Code",
          "example": 200
        },
        "data": {
          "$ref": "#/definitions/AuthSuccessData"
        },
        "message": {
          "description": "response message",
          "type": "string",
          "x-go-name": "Msg",
          "example": "ok"
        }
      },
      "x-go-package": "github.com/jackmrzhou/gc-ai-backend/routers/api/json-models"
    },
    "AuthSuccessData": {
      "type": "object",
      "properties": {
        "token": {
          "description": "jwt token",
          "type": "string",
          "x-go-name": "Token"
        }
      },
      "x-go-package": "github.com/jackmrzhou/gc-ai-backend/routers/api/json-models"
    },
    "Battle": {
      "type": "object",
      "properties": {
        "attacker_id": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "AttackerID"
        },
        "defender_id": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "DefenderID"
        },
        "detail": {
          "type": "string",
          "x-go-name": "Detail"
        },
        "game_id": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "GameID"
        },
        "penalty_score": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "PenaltyScore"
        },
        "reward_score": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "RewardScore"
        },
        "status": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "Status"
        },
        "winner_id": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "WinnerID"
        }
      },
      "x-go-package": "github.com/jackmrzhou/gc-ai-backend/models"
    },
    "BriefBattleInfo": {
      "type": "object",
      "properties": {
        "attacker_id": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "AttackerID"
        },
        "battle_id": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "BattleID"
        },
        "defender_id": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "DefenderID"
        },
        "game_id": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "GameID"
        }
      },
      "x-go-package": "github.com/jackmrzhou/gc-ai-backend/routers/api/json-models"
    },
...
    "GetSourceCodes": {
      "type": "object",
      "properties": {
        "code": {
          "description": "response status",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Code",
          "example": 200
        },
        "data": {
          "description": "source codes",
          "type": "array",
          "items": {
            "$ref": "#/definitions/SourceCode"
          },
          "x-go-name": "Data",
          "example": [
            {
              "code_type": 1,
              "game_id": 2,
              "language": "cpp",
              "source_code": "string",
              "user_id": 1
            }
          ]
        },
        "message": {
          "description": "response message",
          "type": "string",
          "x-go-name": "Msg",
          "example": "ok"
        }
      },
      "x-go-package": "github.com/jackmrzhou/gc-ai-backend/routers/api/json-models"
    },
    "GetUserBattlesResp": {
      "type": "object",
      "properties": {
        "code": {
          "description": "response status",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Code",
          "example": 200
        },
        "data": {
          "description": "battles",
          "type": "array",
          "items": {
            "$ref": "#/definitions/BriefBattleInfo"
          },
          "x-go-name": "Data",
          "example": [
            {
              "attacker_id": 13,
              "battle_id": 22,
              "defender_id": 14,
              "game_id": 15
            }
          ]
        },
        "message": {
          "description": "response message",
          "type": "string",
          "x-go-name": "Msg",
          "example": "ok"
        }
      },
      "x-go-package": "github.com/jackmrzhou/gc-ai-backend/routers/api/json-models"
    },
    "JSONData": {
      "type": "object",
      "x-go-package": "github.com/jackmrzhou/gc-ai-backend/routers/api/json-models"
    },
    "JSONTemplate": {
      "type": "object",
      "properties": {
        "code": {
          "description": "response code",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Code"
        },
        "data": {
          "$ref": "#/definitions/JSONData"
        },
        "message": {
          "description": "response message contains status message",
          "type": "string",
          "x-go-name": "Msg"
        }
      },
      "x-go-package": "github.com/jackmrzhou/gc-ai-backend/routers/api/json-models"
    },
...
    "QueryProcessResp": {
      "type": "object",
      "properties": {
        "code": {
          "description": "response status",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Code",
          "example": 200
        },
        "data": {
          "$ref": "#/definitions/Battle"
        },
        "message": {
          "description": "response message",
          "type": "string",
          "x-go-name": "Msg",
          "example": "ok"
        }
      },
      "x-go-package": "github.com/jackmrzhou/gc-ai-backend/routers/api/json-models"
    },
    "RegInfo": {
      "type": "object",
      "required": [
        "email",
        "password",
        "code"
      ],
      "properties": {
        "code": {
          "type": "string",
          "x-go-name": "VeriCode",
          "example": "778899"
        },
        "email": {
          "description": "registration information",
          "type": "string",
          "x-go-name": "Email",
          "example": "test@email.com"
        },
        "password": {
          "type": "string",
          "x-go-name": "Password",
          "example": "123456"
        }
      },
      "x-go-package": "github.com/jackmrzhou/gc-ai-backend/routers/api/json-models"
    },
    "SourceCode": {
      "type": "object",
      "properties": {
        "code_type": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "CodeType"
        },
        "game_id": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "GameID"
        },
        "language": {
          "type": "string",
          "x-go-name": "Language"
        },
        "source_code": {
          "type": "string",
          "x-go-name": "Content"
        },
        "user_id": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "UserID"
        }
      },
      "x-go-package": "github.com/jackmrzhou/gc-ai-backend/models"
    },
    "StartBattleData": {
      "type": "object",
      "properties": {
        "battle_id": {
          "description": "battle id",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "BattleID",
          "example": 13
        },
        "status": {
          "description": "Current Battle Status: 1 == suspending, 2 == judging, 3 == finished",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "Status",
          "example": 2
        }
      },
      "x-go-package": "github.com/jackmrzhou/gc-ai-backend/routers/api/json-models"
    },
    "StartBattleReq": {
      "type": "object",
      "required": [
        "game_id",
        "user_id"
      ],
      "properties": {
        "game_id": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "GameID",
          "example": 11
        },
        "user_id": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "UserID",
          "example": 15
        }
      },
      "x-go-package": "github.com/jackmrzhou/gc-ai-backend/routers/api/json-models"
    },
    "StartBattleResp": {
      "type": "object",
      "properties": {
        "code": {
          "description": "response code",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Code",
          "example": 200
        },
        "data": {
          "$ref": "#/definitions/StartBattleData"
        },
        "message": {
          "description": "response status",
          "type": "string",
          "x-go-name": "Msg",
          "example": "ok"
        }
      },
      "x-go-package": "github.com/jackmrzhou/gc-ai-backend/routers/api/json-models"
    },
    "TargetEmail": {
      "type": "object",
      "required": [
        "email"
      ],
      "properties": {
        "email": {
          "type": "string",
          "x-go-name": "Email",
          "example": "test@email.com"
        }
      },
      "x-go-package": "github.com/jackmrzhou/gc-ai-backend/routers/api/json-models"
    },
    "UploadSourceCodeData": {
      "type": "object",
      "properties": {
        "source_code_id": {
          "description": "source code's ID",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "SourceCodeID",
          "example": 23
        }
      },
      "x-go-package": "github.com/jackmrzhou/gc-ai-backend/routers/api/json-models"
    },
    "UploadSourceCodeReq": {
      "type": "object",
      "required": [
        "game_id",
        "code_type",
        "language",
        "source_code"
      ],
      "properties": {
        "code_type": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "CodeType",
          "example": 1
        },
        "game_id": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "GameID",
          "example": 12
        },
        "language": {
          "type": "string",
          "x-go-name": "Language",
          "example": "cpp"
        },
        "source_code": {
          "type": "string",
          "x-go-name": "SourceCode"
        }
      },
      "x-go-package": "github.com/jackmrzhou/gc-ai-backend/routers/api/json-models"
    },
    "UploadSourceCodeResp": {
      "type": "object",
      "properties": {
        "code": {
          "description": "response status",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Code",
          "example": 200
        },
        "data": {
          "$ref": "#/definitions/UploadSourceCodeData"
        },
        "message": {
          "description": "response message",
          "type": "string",
          "x-go-name": "Msg",
          "example": "ok"
        }
      },
      "x-go-package": "github.com/jackmrzhou/gc-ai-backend/routers/api/json-models"
    },
    "UserJSON": {
      "type": "object",
      "required": [
        "email",
        "password"
      ],
      "properties": {
        "email": {
          "type": "string",
          "x-go-name": "Email",
          "example": "test@email.com"
        },
        "password": {
          "type": "string",
          "x-go-name": "Password",
          "example": "123456"
        }
      },
      "x-go-package": "github.com/jackmrzhou/gc-ai-backend/routers/api/json-models"
    },
...
  },
  "responses": {
...
    "authSuccess": {
      "description": "authentication passed response",
      "schema": {
        "$ref": "#/definitions/AuthSuccess"
      }
    },
...
    "getSourceCodesResp": {
      "description": "contains all the source codes",
      "schema": {
        "$ref": "#/definitions/GetSourceCodes"
      }
    },
    "getUserBattlesResp": {
      "description": "contains all battles of a user",
      "schema": {
        "$ref": "#/definitions/GetUserBattlesResp"
      }
    },
    "queryProcessResp": {
      "description": "query succeeded, the details of a battle will be returned",
      "schema": {
        "$ref": "#/definitions/QueryProcessResp"
      }
    },
    "startBattleResp": {
      "description": "operation succeeded, the info of the battle will be returned",
      "schema": {
        "$ref": "#/definitions/StartBattleResp"
      }
    },
    "statusResponse": {
      "description": "generic response",
      "schema": {
        "$ref": "#/definitions/JSONTemplate"
      }
    },
    "uploadSourceCodeResp": {
      "description": "operation succeeded, the source code's ID will returned",
      "schema": {
        "$ref": "#/definitions/UploadSourceCodeResp"
      }
    },
...
  }
}